requirements.txt

# Core dependencies
youtube-transcript-api
requests
pandas
streamlit
python-dotenv
google-generativeai